editwindow: allow zooming up to 0.5 tick per pixel

+++ src/gui/dataroll.cpp
@@ -234,7 +234,7 @@ DataRoll::on_draw(const Cairo::RefPtr<Cairo::Context>& cr)
 
 
 void
-DataRoll::set_zoom(int zoom)
+DataRoll::set_zoom(double zoom)
 {
     if (zoom < c_min_zoom) zoom = c_min_zoom;
     else if (zoom > c_max_zoom) zoom = c_max_zoom;
