+++ src/gui/timeroll.cpp
@@ -66,7 +66,7 @@ TimeRoll::on_draw(const Cairo::RefPtr<Cairo::Context>& cr)
 
     int ticks_per_measure =  m_sequence->get_bpm() * (4 * c_ppqn) / m_sequence->get_bw();
     int ticks_per_m_line =  ticks_per_measure * measures_per_line;
-    int start_tick = max(m_hscroll - c_keys_width * m_zoom, 0);
+    int start_tick = max(int(m_hscroll - c_keys_width * m_zoom), 0);
     int end_tick = start_tick + width * m_zoom;
     if (m_sequence->get_length() < end_tick) end_tick = m_sequence->get_length();
     int m = ceil(1.0 * start_tick / ticks_per_measure);
@@ -97,7 +97,7 @@ TimeRoll::on_draw(const Cairo::RefPtr<Cairo::Context>& cr)
 }
 
 void
-TimeRoll::set_zoom(int zoom)
+TimeRoll::set_zoom(double zoom)
 {
     if (zoom < c_min_zoom) zoom = c_min_zoom;
     else if (zoom > c_max_zoom) zoom = c_max_zoom;
